+++ srcs/textures.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   textures.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: david <david@student.42.fr>                +#+  +:+       +#+        */
+/*   By: sbrella <sbrella@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/02 15:55:58 by sbrella           #+#    #+#             */
-/*   Updated: 2019/10/02 01:18:16 by david            ###   ########.fr       */
+/*   Updated: 2019/10/21 18:34:33 by sbrella          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ void					get_texture(char *name, t_txture *texture)
 	texture->width = surf->w;
 	texture->height = surf->h;
 	if (texture->height * texture->width <= 4096 * 2048)
-		ft_memcpy(texture->texture, surf->pixels, (surf->h) * surf->pitch);
+		ft_memcpy(texture->texture, surf->pixels, (size_t)((surf->h) * surf->pitch));
 	else
 		exit(0);
 	// SDL_FreeSurface(surf);
